pyproject.toml
[tool.poetry]
name = "importer-generator"
version = "0.1.0"
description = ""
authors = ["Nick Moreton <nickmoreton@me.com>"]
readme = "README.md"
packages = [{include = "importer_generator"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
requests = "^2.28.2"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
black = "^22.12.0"
flake8 = "^6.0.0"
isort = "^5.11.4"
responses = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
cli = "importer_generator.cli:cli"
